Fix main

+++ main.go
@@ -21,7 +21,7 @@ func init() {
 func main() {
 	// Create the logger.
 	l := logger.NewLogger("ambient", "1.0")
-	l.SetLevel(logrus.DebugLevel)
+	l.SetLevel(uint32(logrus.DebugLevel))
 	// l.SetLevel(logrus.InfoLevel)
 	// l.SetLevel(logrus.ErrorLevel)
 	// l.SetLevel(logrus.FatalLevel)
